:recycle: user 서비스 레이어 아키텍쳐 변경
