+++ src/main/kotlin/team/themoment/gsmNetworking/domain/mentee/service/MenteeRegistrationService.kt
@@ -5,11 +5,11 @@ import team.themoment.gsmNetworking.domain.mentee.domain.Mentee
 import team.themoment.gsmNetworking.domain.mentee.dto.MenteeRegistrationDto
 import team.themoment.gsmNetworking.domain.mentee.repository.MenteeRepository
 import team.themoment.gsmNetworking.domain.user.dto.UserRegistrationDto
-import team.themoment.gsmNetworking.domain.user.service.UserRegistrationService
+import team.themoment.gsmNetworking.domain.user.service.GenerateUserUseCase
 
 @Service
 class MenteeRegistrationService(
-    private val userRegistrationService: UserRegistrationService,
+    private val generateUserUseCase: GenerateUserUseCase,
     private val menteeRepository: MenteeRepository
 ) {
 
@@ -22,7 +22,7 @@ class MenteeRegistrationService(
             null,
             menteeRegistrationDto.profileUrl
         )
-        val user = userRegistrationService.execute(userRegistrationDto, authenticationId)
+        val user = generateUserUseCase.generateUser(userRegistrationDto, authenticationId)
         val mentee = Mentee(user)
 
         menteeRepository.save(mentee)
