+++ src/main/kotlin/team/themoment/gsmNetworking/domain/user/service/GenerateProfileUrlUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.user.service
+
+import team.themoment.gsmNetworking.domain.user.dto.ProfileUrlRegistrationDto
+
+interface GenerateProfileUrlUseCase {
+
+    fun generateProfileUrl(authenticationId: Long, dto: ProfileUrlRegistrationDto)
+}
