+++ src/main/kotlin/team/themoment/gsmNetworking/domain/user/service/ModifyMyUserInfoUseCase.kt
@@ -1,0 +1,8 @@
+package team.themoment.gsmNetworking.domain.user.service
+
+import team.themoment.gsmNetworking.domain.user.dto.UserUpdateInfoDto
+
+interface ModifyMyUserInfoUseCase {
+
+    fun modifyMyUserInfo(authenticationId: Long, userUpdateInfoDto: UserUpdateInfoDto)
+}
